Move unihan to directory module
