+++ cihai/core.py
@@ -9,9 +9,9 @@ import kaptan
 from appdirs import AppDirs
 
 from cihai import exc, extend
-from cihai.db import Database
 from cihai.config import expand_config
 from cihai.constants import DEFAULT_CONFIG
+from cihai.db import Database
 from cihai.utils import import_string, merge_dict
 
 from ._compat import string_types
