+++ examples/basic_usage.py
@@ -3,13 +3,13 @@
 
 from __future__ import print_function, unicode_literals
 
-from cihai.bootstrap import bootstrap_unihan
 from cihai.core import Cihai
+from cihai.unihan.bootstrap import bootstrap_unihan
 
 
 def run(unihan_options={}):
     c = Cihai()
-    c.add_dataset('cihai.unihan.Unihan', namespace='unihan')
+    c.add_dataset('cihai.unihan.dataset.Unihan', namespace='unihan')
 
     if not c.sql.is_bootstrapped:  # download and install Unihan to db
         bootstrap_unihan(c.sql.metadata, options=unihan_options)
